requirements.txt
fastapi[standard]==0.115.5
uvicorn==0.32.1
stomp.py==8.2.0
pydantic==2.10.2
python-multipart==0.0.17
psutil==6.1.0
SQLAlchemy==2.0.36
sqlalchemy-stubs==0.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.8
semver==3.0.4
pyyaml==6.0.2
patool==4.0.0
kubernetes==32.0.1
fastapi-proxy-lib==0.3.0
pipe==2.2